+++ pkg/api/dashboard_snapshot.go
@@ -66,6 +66,7 @@ func GetDashboardSnapshot(c *middleware.Context) {
 	dto := dtos.DashboardFullWithMeta{
 		Dashboard: snapshot.Dashboard,
 		Meta: dtos.DashboardMeta{
+			Type:       m.DashTypeSnapshot,
 			IsSnapshot: true,
 			Created:    snapshot.Created,
 			Expires:    snapshot.Expires,